rename folder examples -> example. improvements to design documents
