+++ example/serverCH/keys/zonePrivate.key
@@ -1,0 +1,1 @@
+b26b4dc83b98f8ceef291b0d5e4c9a7fe7795dac4dc3be7af5a62f5f18fbab8d9d6922d9268c1a32ada9da27a57d4f3e3d093a58c9533df71e5abe5e350b8cd5
