--- examples/serverCH/keys/zonePublic.key
@@ -1,1 +1,0 @@
-9d6922d9268c1a32ada9da27a57d4f3e3d093a58c9533df71e5abe5e350b8cd5
