--- examples/serverRoot/config/server.crt
@@ -1,23 +1,0 @@
------BEGIN CERTIFICATE-----
-MIID5TCCAs2gAwIBAgIJAJGmPmx+xCpcMA0GCSqGSIb3DQEBCwUAMIGIMQswCQYD
-VQQGEwJDSDEPMA0GA1UECAwGWnVyaWNoMQ8wDQYDVQQHDAZadXJpY2gxDDAKBgNV
-BAoMA0VUSDEPMA0GA1UECwwGTmV0U2VjMQ8wDQYDVQQDDAZzZXJ2ZXIxJzAlBgkq
-hkiG9w0BCQEWGGZlaGxtYWNoQHN0dWRlbnQuZXRoei5jaDAeFw0xNzAzMTMxMDE3
-MTlaFw0yNzAzMTExMDE3MTlaMIGIMQswCQYDVQQGEwJDSDEPMA0GA1UECAwGWnVy
-aWNoMQ8wDQYDVQQHDAZadXJpY2gxDDAKBgNVBAoMA0VUSDEPMA0GA1UECwwGTmV0
-U2VjMQ8wDQYDVQQDDAZzZXJ2ZXIxJzAlBgkqhkiG9w0BCQEWGGZlaGxtYWNoQHN0
-dWRlbnQuZXRoei5jaDCCASIwDQYJKoZIhvcNAQEBBQADggEPADCCAQoCggEBANaI
-rGRyjAKsj7Ma1xwXT7vlB1Bstr60GRQr/QKCr3EadhgXfX4u4SlCnoswVV8X8t9v
-Ik8lxozfO/jYrJd/J2YG+cRHlm5dVR2t3cAe+AhhOyMi9tpBcGY9uizeD5sPyGqh
-3ZF4XjqMQyN5N8OANNNCEs87zYfwVDzifR2tYpZdUMKjI0G7WpydSmywKjZq11VE
-8rgd6vOGimLOLZaxS3yA+N2d8L9YAohBAKrhCUaDnqt5Yj092e3QP5hBuyjR+NSd
-vCVX/fAMVMkUSD+QpLR8RYEK8ykHCZzWJaNO6vH41KAyZiE34H4rg05booADnF0B
-gbDY2ClVV/iwYs0KgIkCAwEAAaNQME4wHQYDVR0OBBYEFAkWeBV9SFceJAxe6J/g
-ZeC2fJeZMB8GA1UdIwQYMBaAFAkWeBV9SFceJAxe6J/gZeC2fJeZMAwGA1UdEwQF
-MAMBAf8wDQYJKoZIhvcNAQELBQADggEBAGSKwxsOau6GcQEF7La3aoVba3bRanQh
-/EJVzSc4GECTgonMFnn3PfzOTTG4iL5FPyLZ9Hu3pJXwP7eyHwR9sGYvvepDhwXA
-0syIuR282H06ByXwl8nIQRjRi1agISEZAyp1Y3iEkEjmCE1PUKAK4qzFvSTSdJJv
-tBk5pNPDR/UJwH1kK375cpeFjSH4sw4yIz13fAfrOV2y5n7yN8/dj1pNse7V5vKo
-thj2gY5vhK5JpSdRP5Tiwb6nju/zj8AxxVpWlX3I6PeQ+yCTPPVvIBCd1EiJwYyI
-y+QQwhgbAkl8kkGwLh5q8TcgFeHzkHcc/nQQdoMRBBnGpjKZ44Egrpg=
------END CERTIFICATE-----
