--- examples/serverRootPub/config/rainspub.conf
@@ -1,14 +1,0 @@
-{
-"AssertionValidSince": 0,
-"DelegationValidSince": 0,
-"ShardValidSince": 0,
-"ZoneValidSince": 0,
-"AssertionValidUntil": 86400,
-"DelegationValidUntil": 86400,
-"ShardValidUntil": 86400,
-"ZoneValidUntil": 86400,
-"MaxAssertionsPerShard": 5,
-"ServerAddresses": [{"Type":"TCP", "TCPAddr":{"IP":"127.0.0.1","Port":5023,"Zone":""}}],
-"ZoneFilePath":"zoneFiles/rootZoneFile.txt",
-"ZonePrivateKeyPath": "keys/rootPrivate.key"
-}
