--- examples/serverRootPub/keys/rootPrivate.key
@@ -1,1 +1,0 @@
-f6a5cf1667fdd9949e768561d360c85ed7f7b5a89a9ee6a7bc17c24b18dd5d11db54208448bc95477aeab27d99c23bf70af7053520dd452eca0199de4879a683
