--- examples/serverRootPub/zoneFiles/rootZoneFile.txt
@@ -1,5 +1,0 @@
-:Z: . .   [
-                :S: [
-                        :A: ch [ :deleg: ed25519 2245137bcb058f799843bb8c6df31927b547e4951142b99ae97c668b076e9d84 ]
-                    ]
-            ]
